playing with linear solvers;
seems that LSMR on the Levenberg-Marquadt
system is a bit faster than CG Newton on the
squared system. Sadly have not been able yet
to warm-start reliably the linear solve, perhaps
the line search interacts with that, when we make very small
steps.
It is very tricky to do anything more sophisticated than a
backtracking line search (which however needs to be optimized
to avoid many unnecessary evals). What I want to try next
is to solve a Woodbury simplification of the system, which
is also tricky because the system is not full rank.